Moving endpoints to module
